Changed resolves
As per issue #60 submitted by mark0978 I changed all resolves in the
service calls to one "data" parameter and the injected parameter
listings of the controllers to just a "data" parameter.   Hopefully this
will help with being able to decorate each service and still maintain
the use of the original controller.